decouple AbstractIsAbleFunCallRector
